Fix 'nil' error; add 5th param-result of GetSendMailItem
Something in the data for my client makes the mod try and add a blank line to the history window.  It creates a blank row; but whatever it was passing to FormatMoneyToString would generate an error because it wasn't a number.  So force it to be a number or 0.
As of 7.0.3, GetSendMailItem returns 5 values.